Reorganize project
This commit addresses the issue that makes it impossible
to properly release the product.
The root cause of the issue is the self-reference inside
yaml configuration, which requires to use of a hash/tag
that doesn't exist before the commit is merged.
The only way to solve this - is to remove self-reference,
and to consider the yaml configuration as the part of
the product, not as a separate entity.

So, this commit:
- removes self-reference from the yaml;
- moves yocto layers into the dedicated directory;
- adds .gitignore with the list of build artifacts;
- updates the doc/building.md.

Also, this commit adds changes that are out of the scope
for the commit, like adding info to the README.md regarding
fetching. But this is required part of the change due to
missing parts of the documentation that is important
to properly fetch and build the product.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>